move get image function int sp_global_def
